Update OpenBSD from 5.7 to 6.0
Add a permanent link layer “route” for 180.181.244.1, so that the
routes for 180.181.244.0/23 and 180.181.246.0/23 (both of which have
180.181.244.1 as their destination) don’t “shadow” the very address
through which they need to be sent (and break the Internet as well).

For some reason, this wasn’t necessary when I was on OpenBSD 5.7, but
while I’m not sure whether or not the upgrade really broke it, my best
guess is that it only used to work as a result of a race condition.

“::/0” no longer seems to substitute for “-inet6 default”, but I don’t
know why. See <E1bzNQH-00085z-Kb@uranus.daz.cat> on <misc@openbsd.org>
for possible answers.