dust her off now that we’ve moved out together
We’re now on another SkyMesh NBN FTTP service, but this commit skips
Alkimos (where daria only ran for a month or so) and Arncliffe (where
daria sat on a shelf for over 18 months).

We’re no longer on the same subnet as chi.kingking9.com, so the old
indirect routing hack (afae25bd0eb19) can now be removed.